fix: remove all the mongodb models

+++ organizations/organizations.controllers.js
@@ -144,6 +144,17 @@ const deleteOrganization = asyncHandler(async (req, res) => {
   const { id } = req.params;
 
   try {
+    // Check if the organization with the specified ID exists
+    const existingOrganization = await prisma.organization.findUnique({
+      where: { id },
+    });
+
+    if (!existingOrganization) {
+      return res
+        .status(404)
+        .json({ error: `Organization with ID ${id} not found.` });
+    }
+
     await prisma.organization.delete({
       where: { id },
     });
@@ -156,8 +167,8 @@ const deleteOrganization = asyncHandler(async (req, res) => {
 
 export {
   createOrganization,
+  deleteOrganization,
   getAllOrganizations,
   getOrganizationById,
   updateOrganization,
-  deleteOrganization,
 };
