+++ users/users.controllers.js
@@ -15,6 +15,12 @@ const getSingleUser = asyncHandler(async (req, res) => {
     const user = await prisma.user.findUnique({
       where: { id: id },
     });
+
+    if (!user) {
+      return res
+        .status(404)
+        .json({ message: `User with the specified ID ${id} was not found` });
+    }
     res.status(200).json(user);
   } catch (err) {
     throw new Error(err);
@@ -33,6 +39,15 @@ const getAllUsers = asyncHandler(async (req, res) => {
 const deleteUser = asyncHandler(async (req, res) => {
   const { id } = req.params;
   try {
+    const user = await prisma.user.findFirst({
+      where: { id: id },
+    });
+
+    if (!user) {
+      return res
+        .status(404)
+        .json({ message: `User with the specified ID ${id} was not found` });
+    }
     const deletedUser = await prisma.user.delete({
       where: {
         id,
@@ -51,24 +66,4 @@ const deleteUser = asyncHandler(async (req, res) => {
   }
 });
 
-const updateUser = asyncHandler(async (req, res) => {
-  const { id } = req.params;
-  const { name, phone } = req.body;
-  try {
-    const updatedUser = await User.findByIdAndUpdate(
-      id,
-      { $set: { name, phone } },
-      { new: true }
-    );
-
-    if (!updatedUser) {
-      return res
-        .status(404)
-        .json({ message: `User with the specified ID ${id} was not found` });
-    }
-    res.status(202).json(updatedUser);
-  } catch (err) {
-    throw new Error(err);
-  }
-});
-export { getSingleUser, getAllUsers, updateUser, deleteUser };
+export { deleteUser, getAllUsers, getSingleUser };
